TDD/Money.cs
namespace TDD
{
    public class Money : IExpression
    {
        protected int _amount;
        protected string _currency;

        public Money(int amount, string currency)
        {
            _amount = amount;
            _currency = currency;
        }

        public int Amount { get { return _amount; } }
        public string Currency { get { return _currency; } }

        public IExpression Times(int multiplier)
        {
            return new Money(multiplier * _amount, _currency);
        }

        public IExpression Plus(IExpression addend)
        {
            return new Sum(this, addend);
        }

        public Money Reduce(Bank bank, string to)
        {
            int rate = bank.Rate(_currency, to);
            return new Money(_amount / rate, to);
        }

        public override Boolean Equals(Object? obj)
        {
            if (obj is Money money)
            {
                return _amount == money.Amount && _currency.Equals(money.Currency);
            }
            return false;
        }

        public override int GetHashCode()
        {
            return _amount.GetHashCode();
        }

        public static Money Dollar(int amount)
        {
            return new Money(amount, "USD");
        }

        public static Money Franc(int amount)
        {
            return new Money(amount, "CHF");
        }
    }
}
